+++ go/api/connection_management_commands_test.go
@@ -5,6 +5,7 @@ package api
 import (
 	"fmt"
 
+	"github.com/google/uuid"
 	"github.com/valkey-io/valkey-glide/go/api/options"
 )
 
@@ -50,6 +51,30 @@ func ExampleGlideClient_ClientId() {
 	}
 	assert := result > 0
 	fmt.Println(assert)
+
+	// Output: true
+}
+
+func ExampleGlideClient_ClientSetName() {
+	var client *GlideClient = getExampleGlideClient() // example helper function
+	result, err := client.ClientSetName("ConnectionName")
+	if err != nil {
+		fmt.Println("Glide example failed with an error: ", err)
+	}
+	fmt.Println(result)
+
+	// Output: OK
+}
+
+func ExampleGlideClient_ClientGetName() {
+	var client *GlideClient = getExampleGlideClient() // example helper function
+	connectionName := "ConnectionName-" + uuid.NewString()
+	client.ClientSetName(connectionName)
+	result, err := client.ClientGetName()
+	if err != nil {
+		fmt.Println("Glide example failed with an error: ", err)
+	}
+	fmt.Println(result == connectionName)
 
 	// Output: true
 }