+++ go/api/server_management_cluster_commands.go
@@ -37,4 +37,12 @@ type ServerManagementClusterCommands interface {
 	ConfigResetStat() (string, error)
 
 	ConfigResetStatWithOptions(routeOption options.RouteOption) (string, error)
+
+	ConfigSet(parameters map[string]string) (string, error)
+
+	ConfigSetWithOptions(parameters map[string]string, routeOption options.RouteOption) (string, error)
+
+	ConfigGet(parameters []string) (ClusterValue[interface{}], error)
+
+	ConfigGetWithOptions(parameters []string, routeOption options.RouteOption) (ClusterValue[interface{}], error)
 }